Change how userdata works
This previously would always include a puppet MCO specific script in the userdata. This change makes it not include that by default. The new variable "bastion_userdata_template" can be set to "puppet_masterless_mco" to enable it again.

There is also a slight change to how custom userdata is combined with the puppet-specific script. This will now be included AFTER the bash shebang. And the bash path has been changed for portability.