Add ICMP support
Adding ICMP support so that you can allow CIDRs to ping the Bastion.
This is being added because Superdry need to be able to ping the bastion
from their monitoring server in Windcheater as their VPN keeps dropping
because there is insufficient traffic.

This is documented here:

https://aws.amazon.com/premiumsupport/knowledge-center/vpn-tunnel-instability-inactivity/

It is set by default to allow ICMP traffic from everywhere.